Initial commit, set up docker repos for lambda dispatch/batch calls
