--- .dockerignore
@@ -1,6 +1,0 @@
-data/job_queries
-data/fasttext
-data/kaggle_postings
-env/
-workspace.ipynb
-job_scrape.ipynb
