+++ .gitignore
@@ -1,6 +1,4 @@
 /env
-/data/fasttext
-/data/kaggle_postings
-/data/job_queries
-/src/proxies.py
+/data
+*proxies.py
 /models
