+++ docker_analyzer/analysis.py
