+++ docker_analyzer/data/countries.txt
